config: add TOML config parser
Add a basic parser for TOML config file with some of the first options we might
support getting from a config.  It's still a work in progress: this patch only
check if the topics and options placed in the config file (~/.tbott.toml) are
valid; the code doesn't populate any structure yet.