treewide: major project rework
This patch is most likely a patch bomb, where many different things are changed
with no direct relation between them.  This patch not only rename the project
(AGAIN) from CBot to BotZao, but also change the whole architecture, which,
now, is more sane for supporting different IM protocols/applications (IRC,
Telegram, Discord, ...) in the near future.

This is an example of BAD COMMIT.  It MUST NOT happen again.