plugins: rework return value logic
Perl's 'die' call watch for undefined values and any value, including 0
is a defined int value, so it's not triggered when it's returned. With
that in mind, subroutines that have a real return value (non void
context) should return 'undef' in case of simple errors and, unlike some
languages like C where 0 is used for success, return 1 for in case of
success, like done when calling a module (the 1; at the very end of the
file).

At the same time, remove explicit `return` calls from those subs that
are used in void context only (no return value).